feat: enhance AMQP channel and connection handling with new frame creation methods
