+++ internal/core/amqp/channel.go
@@ -1,11 +1,30 @@
 package amqp
 
-func closeChannelFrame(request *RequestMethodMessage) []byte {
+func closeChannelFrame(channel, replycode, classID, methodID uint16, replyText string) []byte {
+	replyCodeKv := KeyValue{
+		Key:   INT_SHORT,
+		Value: replycode,
+	}
+	replyTextKv := KeyValue{
+		Key:   STRING_SHORT,
+		Value: replyText,
+	}
+	classIDKv := KeyValue{
+		Key:   INT_SHORT,
+		Value: classID,
+	}
+	methodIDKv := KeyValue{
+		Key:   INT_SHORT,
+		Value: methodID,
+	}
+	content := ContentList{
+		KeyValuePairs: []KeyValue{replyCodeKv, replyTextKv, classIDKv, methodIDKv},
+	}
 	frame := ResponseMethodMessage{
-		Channel:  request.Channel,
+		Channel:  channel,
 		ClassID:  uint16(CHANNEL),
 		MethodID: uint16(CHANNEL_CLOSE_OK),
-		Content:  ContentList{},
+		Content:  content,
 	}.FormatMethodFrame()
 	return frame
 }
@@ -23,6 +42,16 @@ func createChannelOpenOkFrame(request *RequestMethodMessage) []byte {
 				},
 			},
 		},
+	}.FormatMethodFrame()
+	return frame
+}
+
+func createChannelCloseOkFrame(channel uint16) []byte {
+	frame := ResponseMethodMessage{
+		Channel:  channel,
+		ClassID:  uint16(CHANNEL),
+		MethodID: uint16(CHANNEL_CLOSE_OK),
+		Content:  ContentList{},
 	}.FormatMethodFrame()
 	return frame
 }
